+++ Vimari Extension/js/link-hints.js
@@ -354,7 +354,7 @@ function numberToHintString(number, numHintDigits) {
 function simulateClick(link, openInNewTab) {
   if (openInNewTab) {
     console.log("-- Open link in new tab --");
-    safari.extension.dispatchMessage("openLinkInTab", { url: link.href });
+    extensionCommunicator.requestOpenLinkInBackground(link.href);
   } else {
     link.click();
   }